+++ lib/src/painters/expanding_dots_painter.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:smooth_page_indicator/src/effects/expanding_dots_effect.dart';
+import '../effects/expanding_dots_effect.dart';
 
 import 'indicator_painter.dart';
 
@@ -10,22 +10,21 @@ class ExpandingDotsPainter extends IndicatorPainter {
     @required double offset,
     @required this.effect,
     @required int count,
-    @required bool isRTL,
-  }) : super(offset, count, effect, isRTL);
+  }) : super(offset, count, effect);
 
   @override
   void paint(Canvas canvas, Size size) {
-    final int current = offset.floor();
-    double drawingOffset = -effect.spacing;
+    final current = offset.floor();
+    var drawingOffset = -effect.spacing;
     final dotOffset = offset - current;
 
-    for (int i = 0; i < count; i++) {
-      Color color = effect.dotColor;
+    for (var i = 0; i < count; i++) {
+      var color = effect.dotColor;
       final activeDotWidth = effect.dotWidth * effect.expansionFactor;
       final expansion =
           (dotOffset / 2 * ((activeDotWidth - effect.dotWidth) / .5));
       final xPos = drawingOffset + effect.spacing;
-      double width = effect.dotWidth;
+      var width = effect.dotWidth;
       if (i == current) {
         color = Color.lerp(effect.activeDotColor, effect.dotColor, dotOffset);
         width = activeDotWidth - expansion;